requirements.txt
flask
# SAP AI SDK - Note: proxy_server.py includes a monkey patch to fix client_type compatibility
# between sap-ai-sdk-gen (gen_ai_hub) and ai_core_sdk 2.6.2
ai_core_sdk>=2.6.2
openai
litellm
sap-ai-sdk-gen[all]>=5.10.0
aioboto3

# Desktop Application Dependencies
pywebview>=4.0
Pillow>=9.0
pystray>=0.19.0  # System tray support

# Production Server
waitress>=2.0

# Packaging
pyinstaller>=6.0
